Rework sprint 3 assignments (#694)
Previously they were:
1. Very repetitive, having the trainees solve the same problem more
   than once which was confusing and not a great use of time.
2. Large. Reviewing the whole PR for the sprint took a long time, so was
   very daunting for reviewers, and led to review backlogs.
3. Hard to track who was doing the stretch, because it was maybe in the
   same PR or just maybe missing from the PR.

Instead:
1. Re-use their implementation for both sets of tests - this emphasises
   also that the point of the exercise is writing tests different ways.
2. Split off the practice-tdd and stretch into their own expected
   separate PRs.
3. Fix up a bunch of inconsistent formatting.
