+++ Sprint-3/1-implement-and-rewrite-tests/rewrite-tests-with-jest/1-get-angle-type.test.js
@@ -1,4 +1,6 @@
-const getAngleType = require("./1-get-angle-type");
+// This statement loads the getAngleType function you wrote in the implement directory.
+// We will use the same function, but write tests for it using Jest in this file.
+const getAngleType = require("../implement/1-get-angle-type");
 
 test("should identify right angle (90°)", () => {
   expect(getAngleType(90)).toEqual("Right angle");
