+++ Sprint-3/1-implement-and-rewrite-tests/rewrite-tests-with-jest/2-is-proper-fraction.test.js
@@ -1,4 +1,6 @@
-const isProperFraction = require("./2-is-proper-fraction");
+// This statement loads the isProperFraction function you wrote in the implement directory.
+// We will use the same function, but write tests for it using Jest in this file.
+const isProperFraction = require("../implement/2-is-proper-fraction");
 
 test("should return true for a proper fraction", () => {
   expect(isProperFraction(2, 3)).toEqual(true);
