+++ Sprint-3/1-implement-and-rewrite-tests/rewrite-tests-with-jest/3-get-card-value.test.js
@@ -1,0 +1,13 @@
+// This statement loads the getCardValue function you wrote in the implement directory.
+// We will use the same function, but write tests for it using Jest in this file.
+const getCardValue = require("../implement/3-get-card-value");
+
+test("should return 11 for Ace of Spades", () => {
+  const aceofSpades = getCardValue("A♠");
+  expect(aceofSpades).toEqual(11);
+});
+
+// Case 2: Handle Number Cards (2-10):
+// Case 3: Handle Face Cards (J, Q, K):
+// Case 4: Handle Ace (A):
+// Case 5: Handle Invalid Cards:
