--- Sprint-3/2-mandatory-rewrite/1-get-angle-type.js
@@ -1,18 +1,0 @@
-function getAngleType(angle) {
-  if (angle === 90) return "Right angle";
-  // replace with your completed function from key-implement
-
-}
-
-
-
-
-
-
-
-
-// Don't get bogged down in this detail
-// Jest uses CommonJS module syntax by default as it's quite old
-// We will upgrade our approach to ES6 modules in the next course module, so for now 
-// we have just written the CommonJS module.exports syntax for you
-module.exports = getAngleType;
