--- Sprint-3/2-mandatory-rewrite/2-is-proper-fraction.js
@@ -1,6 +1,0 @@
-function isProperFraction(numerator, denominator) {
-    if (numerator < denominator) return true;
-    // add your completed function from key-implement here
-}
-
-module.exports = isProperFraction;
