--- Sprint-3/2-mandatory-rewrite/3-get-card-value.js
@@ -1,5 +1,0 @@
-function getCardValue(card) {
-    // replace with your code from key-implement
-    return 11;
-}
-module.exports = getCardValue;
