--- Sprint-3/2-mandatory-rewrite/3-get-card-value.test.js
@@ -1,11 +1,0 @@
-const getCardValue = require("./3-get-card-value");
-
-test("should return 11 for Ace of Spades", () => {
-    const aceofSpades = getCardValue("A♠");
-    expect(aceofSpades).toEqual(11);
-    });
-
-// Case 2: Handle Number Cards (2-10):
-// Case 3: Handle Face Cards (J, Q, K):
-// Case 4: Handle Ace (A):
-// Case 5: Handle Invalid Cards:
