+++ Sprint-3/2-practice-tdd/README.md
@@ -1,0 +1,13 @@
+# Practice TDD
+
+In this section you'll practice this key skill of building up your program test first.
+
+Use the Jest syntax and complete the provided files, meeting the acceptance criteria for each function. Use the VSCode test runner to run your tests and check your progress.
+
+Write the tests _before_ the code that will make them pass.
+
+Recommended order:
+
+1. `count.test.js`
+1. `repeat.test.js`
+1. `get-ordinal-number.test.js`
