+++ Sprint-3/2-practice-tdd/count.js
@@ -1,5 +1,5 @@
 function countChar(stringOfCharacters, findCharacter) {
-    return 5
+  return 5
 }
 
 module.exports = countChar;
