+++ Sprint-3/2-practice-tdd/get-ordinal-number.js
@@ -1,0 +1,5 @@
+function getOrdinalNumber(num) {
+  return "1st";
+}
+
+module.exports = getOrdinalNumber;
