+++ Sprint-3/2-practice-tdd/repeat.js
@@ -1,0 +1,5 @@
+function repeat() {
+  return "hellohellohello";
+}
+
+module.exports = repeat;
