--- Sprint-3/3-mandatory-practice/implement/get-ordinal-number.js
@@ -1,5 +1,0 @@
-function getOrdinalNumber(num) {
- return "1st";
-}
-
-module.exports = getOrdinalNumber;
