--- Sprint-3/3-mandatory-practice/implement/repeat.js
@@ -1,5 +1,0 @@
-function repeat() {
-    return "hellohellohello";
-}
-
-module.exports = repeat;
