+++ Sprint-3/3-stretch/README.md
@@ -1,0 +1,9 @@
+# 🔍 Stretch
+
+These stretch activities are not mandatory, but we hope you will explore them after you have completed the mandatory work.
+
+In this exercise, you'll need to **play computer** with the function `find`. This function makes use of while loop statement. Your task will be to step through the code to figure out what is happening when the computer executes the code.
+
+Next, try implementing the functions specified in `password-validator.js`.
+
+Finally, set up your own script and test files for `card-validator.md`
