+++ Sprint-3/3-stretch/password-validator.js
